+++ src/components/Checkout/DeliveryMethods/SelfCollection/index.tsx
@@ -2,20 +2,16 @@ import React, { ChangeEvent, ComponentProps } from 'react';
 import styles from './index.module.scss';
 import { Form, Radio } from 'react-daisyui';
 import { Link } from 'react-router-dom';
-// import dayjs from 'dayjs';
 
 import { CollectionPoint } from '../../../../types/order.type';
-import { ProductType, type Product } from '../../../../types/product.type';
 
 interface Props extends ComponentProps<'div'> {
-  product: Pick<Product, 'type' | 'releasedDate' | 'orderedDate'>;
   data: CollectionPoint;
   onChangeData: (data: CollectionPoint) => void;
 }
 
 export function SelfCollection({
   className,
-  product,
   data,
   onChangeData,
   ...rest
@@ -35,16 +31,6 @@ export function SelfCollection({
       className={`${className} ${styles['delivery-method-choose-collection']}`}
       {...rest}
     >
-      {product.type === ProductType.PreOrder && (
-        <p className={styles['delivery-date']}>
-          {/* *The estimated delivery date of this pre-order is{' '}
-          {dayjs(product.releasedDate).format('DD MMMM YYYY')} */}
-          Your order will be delivered within 2 weeks of the release date and we
-          will contact you to schedule a delivery. Any delays beyond that will
-          be refunded fully.
-        </p>
-      )}
-
       <p>Choose your collection point</p>
 
       <Form className={`${styles['delivery-method-choose-form']}`}>