+++ src/components/Header/index.tsx
@@ -6,6 +6,7 @@ import Logo from '../Logo';
 import List from '../List';
 import Icon from '../Icon';
 import useAuth from '../../hooks/useAuth';
+import useCart from '../../hooks/useCart';
 import responsive from '../../utils/responsive';
 
 interface Props extends ComponentProps<'div'> {}
@@ -13,6 +14,7 @@ interface Props extends ComponentProps<'div'> {}
 function Header({ className, ...rest }: Props) {
   const location = useLocation();
   const { user } = useAuth();
+  const { cartItems } = useCart();
 
   return (
     <div className={`${styles['header']} ${className}`} {...rest}>
@@ -40,20 +42,26 @@ function Header({ className, ...rest }: Props) {
             >
               Contact Us
             </Link>,
-            ...[
-              !user ? (
-                <Link
-                  className={styles['sign-in']}
-                  to={`/auth/login?redirectUrl=${encodeURIComponent(
-                    `${location.pathname}${location.search}`,
-                  )}`}
-                >
-                  Sign In
-                </Link>
-              ) : (
-                <Icon className={styles['icon-user']} name="user" />
-              ),
-            ],
+            !user ? (
+              <Link
+                className={styles['sign-in']}
+                to={`/auth/login?redirectUrl=${encodeURIComponent(
+                  `${location.pathname}${location.search}`,
+                )}`}
+              >
+                Sign In
+              </Link>
+            ) : (
+              <Icon className={styles['icon-user']} name="user" />
+            ),
+            <Link className={styles['shopping-card']} to={'/cart'}>
+              <Icon className={styles['icon-card']} name="shopping-cart" />
+              {cartItems.length > 0 && (
+                <div className={styles['item-number']}>
+                  <span>{cartItems.length}</span>
+                </div>
+              )}
+            </Link>,
           ]}
         />
       </div>