+++ src/utils/transformBEOrder.ts
@@ -3,8 +3,10 @@ import type { Order } from '../types/order.type';
 const transformBEOrder = (
   order: Omit<Order, 'id' | 'paymentStatus' | 'orderedDate'>,
 ): any => ({
-  product_id: order.productId,
-  order_quantity: order.quantity,
+  products: order.products.map((product: any) => ({
+    product_id: product.id,
+    order_quantity: product.quantity,
+  })),
   buyer_id: order.buyerId,
   email: order.contactDetails.email,
   phone_number: order.contactDetails.phoneNumber,
@@ -13,7 +15,6 @@ const transformBEOrder = (
   address_line_2: order.deliveryAddress.addressLine2,
   postal_code: order.deliveryAddress.postalCode,
   fees: {
-    product_price: order.price,
     small_order_fee: order.additionalFee,
     delivery_type: order.deliveryMethod,
     delivery_fee: order.deliveryFee,