+++ src/utils/transformOrder.ts
@@ -2,10 +2,10 @@ import type { Order } from '../types/order.type';
 
 const transformOrder = (order: any): Order => ({
   id: order.order_id || order.guest_order_id,
-  productId: order.product_id,
-  price: order.fees.product_price,
-  quantity: order.order_quantity,
-  subTotal: order.fees.product_price * order.order_quantity,
+  products: order.products.map((product: any) => ({
+    id: product.product_id,
+    quantity: product.order_quantity,
+  })),
   additionalFee: order.fees.small_order_fee,
   deliveryFee: order.fees.delivery_fee,
   paymentFee: order.fees.payment_fee,